Adding roles and store features
